+++ server/neptune/workflows/internal/deploy/revision/queue/worker.go
@@ -1,12 +1,9 @@
 package queue
 
 import (
-	"context"
 	"fmt"
 
-	"github.com/pkg/errors"
 	internalContext "github.com/runatlantis/atlantis/server/neptune/context"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/config/logger"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/terraform"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
@@ -16,21 +13,8 @@ import (
 
 const DeploymentInfoVersion = "1.0.0"
 
-type terraformWorkflowRunner interface {
-	Run(ctx workflow.Context, deploymentInfo terraform.DeploymentInfo) error
-}
-
-type dbActivities interface {
-	FetchLatestDeployment(ctx context.Context, request activities.FetchLatestDeploymentRequest) (activities.FetchLatestDeploymentResponse, error)
-	StoreLatestDeployment(ctx context.Context, request activities.StoreLatestDeploymentRequest) error
-}
-
 type revisionProcessor interface {
-	Process(ctx workflow.Context, requestedDeployment terraform.DeploymentInfo, latestDeployment *root.DeploymentInfo) error
-}
-
-type workerActivities interface {
-	dbActivities
+	Process(ctx workflow.Context, requestedDeployment terraform.DeploymentInfo, latestDeployment *root.DeploymentInfo) (*root.DeploymentInfo, error)
 }
 
 type WorkerState string
@@ -48,10 +32,8 @@ type UnlockSignalRequest struct {
 }
 
 type Worker struct {
-	Queue                   *Queue
-	TerraformWorkflowRunner terraformWorkflowRunner
-	Activities              workerActivities
-	RevisionProcessor       revisionProcessor
+	Queue             *Queue
+	RevisionProcessor revisionProcessor
 
 	// mutable
 	state WorkerState
@@ -91,67 +73,11 @@ func (w *Worker) Work(ctx workflow.Context) {
 
 		ctx := workflow.WithValue(ctx, internalContext.SHAKey, msg.Revision)
 		ctx = workflow.WithValue(ctx, internalContext.DeploymentIDKey, msg.ID)
-
-		latestDeployment, err = w.fetchLatestDeployment(ctx, msg, latestDeployment)
+		latestDeployment, err = w.RevisionProcessor.Process(ctx, msg, latestDeployment)
 		if err != nil {
-			logger.Error(ctx, "unable to fetch latest deployment for root: %s, skipping deploy: %s", msg.Root.Name, err.Error())
-			continue
-		}
-		err = w.RevisionProcessor.Process(ctx, msg, latestDeployment)
-		if err != nil {
-			logger.Error(ctx, "failed to process revision, moving to next one")
-			continue
-		}
-
-		err = w.TerraformWorkflowRunner.Run(ctx, msg)
-		if err != nil {
-			logger.Error(ctx, "failed to deploy revision, moving to next one")
-			continue
-		}
-		latestDeployment = w.buildLatestDeployment(msg)
-
-		// TODO: Persist deployment on shutdown if it fails instead of blocking
-		if err = w.persistLatestDeployment(ctx, latestDeployment); err != nil {
-			logger.Error(ctx, "failed to persist latest deploy job")
+			logger.Error(ctx, "failed to process revision, moving to next one", "err", err)
 		}
 	}
-}
-
-func (w *Worker) buildLatestDeployment(deployRequest terraform.DeploymentInfo) *root.DeploymentInfo {
-	return &root.DeploymentInfo{
-		Version:    DeploymentInfoVersion,
-		ID:         deployRequest.ID.String(),
-		CheckRunID: deployRequest.CheckRunID,
-		Revision:   deployRequest.Revision,
-		Root:       deployRequest.Root,
-		Repo:       deployRequest.Repo,
-	}
-}
-
-func (w *Worker) fetchLatestDeployment(ctx workflow.Context, deploymentInfo terraform.DeploymentInfo, latestDeployment *root.DeploymentInfo) (*root.DeploymentInfo, error) {
-	// Skip fetching latest deployment it it's already in memory
-	if latestDeployment != nil {
-		return latestDeployment, nil
-	}
-	var resp activities.FetchLatestDeploymentResponse
-	err := workflow.ExecuteActivity(ctx, w.Activities.FetchLatestDeployment, activities.FetchLatestDeploymentRequest{
-		FullRepositoryName: deploymentInfo.Repo.GetFullName(),
-		RootName:           deploymentInfo.Root.Name,
-	}).Get(ctx, &resp)
-	if err != nil {
-		return nil, errors.Wrap(err, "fetching latest deployment")
-	}
-	return resp.DeploymentInfo, nil
-}
-
-func (w *Worker) persistLatestDeployment(ctx workflow.Context, deploymentInfo *root.DeploymentInfo) error {
-	err := workflow.ExecuteActivity(ctx, w.Activities.StoreLatestDeployment, activities.StoreLatestDeploymentRequest{
-		DeploymentInfo: deploymentInfo,
-	}).Get(ctx, nil)
-	if err != nil {
-		return errors.Wrap(err, "persisting deployment info")
-	}
-	return nil
 }
 
 func (w *Worker) GetState() WorkerState {
